+++ src/components/index.ts
@@ -3,6 +3,7 @@
  * Note: Please keep all imports, type exports, and component exports sorted alphabetically
  */
 
+import UiBadge from './badge/UiBadge.vue'
 import UiButton from './button/UiButton.vue'
 import UiCarousel from './carousel/UiCarousel.vue'
 import UiCheckbox from './checkbox/UiCheckbox.vue'
@@ -15,6 +16,7 @@ import UiSelect from './select/UiSelect.vue'
 import UiSkeleton from './skeleton/UiSkeleton.vue'
 import UiTooltip from './tooltip/UiTooltip.vue'
 
+export type { UiBadgeProps, UiBadgeSlots } from './badge/UiBadge.vue'
 export type { UiButtonProps, UiButtonEmits, UiButtonSlots } from './button/UiButton.vue'
 export type { UiCarouselProps, UiCarouselEmits, UiCarouselSlots } from './carousel/UiCarousel.vue'
 export type { UiCheckboxProps, UiCheckboxEmits, UiCheckboxSlots } from './checkbox/UiCheckbox.vue'
@@ -29,6 +31,7 @@ export type { UiTooltipProps, UiTooltipSlots } from './tooltip/UiTooltip.vue'
 export type { AppConfig as UiKitConfig } from './types'
 
 export {
+  UiBadge,
   UiButton,
   UiCarousel,
   UiCheckbox,