+++ src/components/types.ts
@@ -1,3 +1,4 @@
+import type { BadgeUi } from './badge/theme'
 import type { ButtonUi } from './button/theme'
 import type { CarouselUi } from './carousel/theme'
 import type { CheckboxUi } from './checkbox/theme'
@@ -19,6 +20,7 @@ export type DeepPartial<T> = {
 export type UiProp<T> = DeepPartial<T> & { strategy?: Strategy }
 
 interface UiConfig {
+  badge?: UiProp<BadgeUi>
   button?: UiProp<ButtonUi>
   carousel?: UiProp<CarouselUi>
   checkbox?: UiProp<CheckboxUi>