+++ src/types/components/index.d.ts
@@ -2,6 +2,7 @@
  * Components exports
  * Note: Please keep all imports, type exports, and component exports sorted alphabetically
  */
+import UiBadge from './badge/UiBadge.vue';
 import UiButton from './button/UiButton.vue';
 import UiCarousel from './carousel/UiCarousel.vue';
 import UiCheckbox from './checkbox/UiCheckbox.vue';
@@ -13,6 +14,7 @@ import UiMenu from './menu/UiMenu.vue';
 import UiSelect from './select/UiSelect.vue';
 import UiSkeleton from './skeleton/UiSkeleton.vue';
 import UiTooltip from './tooltip/UiTooltip.vue';
+export type { UiBadgeProps, UiBadgeSlots } from './badge/UiBadge.vue';
 export type { UiButtonProps, UiButtonEmits, UiButtonSlots } from './button/UiButton.vue';
 export type { UiCarouselProps, UiCarouselEmits, UiCarouselSlots } from './carousel/UiCarousel.vue';
 export type { UiCheckboxProps, UiCheckboxEmits, UiCheckboxSlots } from './checkbox/UiCheckbox.vue';
@@ -25,4 +27,4 @@ export type { UiMenuItem, UiMenuProps } from './menu/UiMenu.vue';
 export type { UiSelectProps, SelectOption } from './select/UiSelect.vue';
 export type { UiTooltipProps, UiTooltipSlots } from './tooltip/UiTooltip.vue';
 export type { AppConfig as UiKitConfig } from './types';
-export { UiButton, UiCarousel, UiCheckbox, UiCircleProgressBar, UiIcon, UiInput, UiLink, UiMenu, UiSelect, UiSkeleton, UiTooltip };
+export { UiBadge, UiButton, UiCarousel, UiCheckbox, UiCircleProgressBar, UiIcon, UiInput, UiLink, UiMenu, UiSelect, UiSkeleton, UiTooltip };