+++ src/types/components/link/UiLink.vue.d.ts
@@ -20,8 +20,8 @@ declare const __VLS_export: __VLS_WithSlots<import("vue").DefineComponent<UiLink
 }, string, import("vue").PublicProps, Readonly<UiLinkProps> & Readonly<{
     onClick?: ((value: MouseEvent) => any) | undefined;
 }>, {
-    ui: UiProp<LinkUi>;
     variant: "primary" | "default";
+    ui: UiProp<LinkUi>;
     to: RouteLocationRaw;
     activeClass: string | false | 0 | 0n | ClassNameValue[] | null;
     exactActiveClass: string | false | 0 | 0n | ClassNameValue[] | null;